Added `Generics` parameters handling to Mocking
In order for Mocks to support generic parameters, needed to create a `struct GenericArgument` that acts as a type erasure type.

One of the goals of these mocks is to have the ability to assert whether they were called with the correct parameters, so we need a way to compare them. There are currently 2 main equality routes:
1. Via an extension on `Matcher` where `T`: `Equatable`.
2. Via an extension on `Matcher` where `T` == `GenericArgument` and the wrapped generic argument is itself `Equatable`

If none of those options applies to a parameter supplied to `Matcher`, then a `fatalError` is raised. Note this is the case only when an `exact` match is required.