Fix `TimestampFormatter` not working with special cases.
While running tests during certain times (i.e. midnight), the logic for the Timestamp formatter was not working.
In order to fix it I'm now leveraging the diffing algorithm from the `Calendar` class, which returns the differences
as `DateComponents` which is what I was looking for.
Additionally broken down the single tests into multiple tests.